scripts/requirements.txt
torch>=1.9.0
transformers>=4.20.0
numpy>=1.21.0
Pillow>=8.3.0